Refactor: Made client num inputs use long long instead of int

+++ senc/client/input.hpp
@@ -14,8 +14,8 @@ namespace senc
 {
 	template <typename Self>
 	concept NumInputable = std::integral<Self> &&
-		(std::numeric_limits<Self>::min() >= std::numeric_limits<int>::min()) &&
-		(std::numeric_limits<Self>::max() <= std::numeric_limits<int>::max());
+		(std::numeric_limits<Self>::min() >= std::numeric_limits<long long>::min()) &&
+		(std::numeric_limits<Self>::max() <= std::numeric_limits<long long>::max());
 
 	std::string input();
 	std::string input(const std::string& msg);
